# perf: improve loading time at startup