--- a/WolvenKit.Common/Services/HashService.cs
+++ b/WolvenKit.Common/Services/HashService.cs
@@ -88,31 +88,31 @@ namespace WolvenKit.Common.Services
         {
             var hashesMemory = DecompressEmbeddedFile(s_used);
             ReadHashes(hashesMemory.GetStream());
-            
+
             hashesMemory.Dispose();
-            
+
             var nodeRefsMemory = DecompressEmbeddedFile(s_nodeRefs);
             ReadNodeRefs(nodeRefsMemory.GetStream());
-            
+
             nodeRefsMemory.Dispose();
-            
+
             var tweakNamesMemory = DecompressEmbeddedFile(s_tweakDbStr);
             ReadTweakNames(tweakNamesMemory.GetStream());
-            
+
             tweakNamesMemory.Dispose();
-            
+
             LoadMissingHashes();
 
             _isLoaded = true;
         }
-        
+
         private static unsafe UnmanagedMemory DecompressEmbeddedFile(string resourceName)
         {
             using var stream = Assembly.GetExecutingAssembly().GetManifestResourceStream(resourceName).NotNull();
 
             // read KARK header
             var oodleCompression = stream.ReadStruct<uint>();
-            
+
             if (oodleCompression != Oodle.KARK)
             {
                 throw new DecompressionException("Incorrect hash file.");
@@ -123,7 +123,7 @@ namespace WolvenKit.Common.Services
             var compressedBufferLength = (int)(stream.Length - (sizeof(uint) * 2));
             using var compressedBuffer = UnmanagedMemory.Allocate(compressedBufferLength);
             var decompressedBuffer = UnmanagedMemory.Allocate((int) outputSize);
-            
+
             // read the rest of the stream
             var read = stream.Read(compressedBuffer.GetSpan());
 
@@ -135,7 +135,7 @@ namespace WolvenKit.Common.Services
             Oodle.Decompress(
                 compressedBuffer.Pointer, compressedBuffer.Size,
                 decompressedBuffer.Pointer, decompressedBuffer.Size);
-            
+
             return decompressedBuffer;
         }
 
@@ -146,7 +146,7 @@ namespace WolvenKit.Common.Services
             var readerTask = Task.Run(() =>
             {
                 using var sr = new StreamReader(memoryStream);
-                
+
                 while (true)
                 {
                     var nextLine = sr.ReadLine();
@@ -158,7 +158,7 @@ namespace WolvenKit.Common.Services
 
                     collection.Add(nextLine);
                 }
-                
+
                 collection.CompleteAdding();
             });