--- a/WolvenKit.Core/Services/IHashService.cs
+++ b/WolvenKit.Core/Services/IHashService.cs
@@ -1,9 +1,14 @@
 using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace WolvenKit.Common.Services
 {
     public interface IHashService
     {
+        bool IsLoaded { get; }
+
+        Task LoadAsync();
+
         bool Contains(ulong key, bool checkUserHashes = true);
 
         string? Get(ulong key);