# perf: improve loading time at startup

**Implemented:**
- asynchronously load hashes (saves ~5 000 ms)
- asynchronously load CRUIDs (saves ~500 ms)
- asynchronously refresh UI scripts (saves ~500 ms)

**Additional notes:**
Tool: dotTrace with timeline profile (+ profile build).
Scenario: timing from the launch to [First Contentful Paint](https://web.dev/articles/fcp) (when Main View is rendered and user can interact).

Elapsed time before: ~20 000 ms
Elapsed time after: ~10 000 ms
Overall improvement: ~50 %

A major remaining bottle neck is during `WolvenKit.Views.Shell.DockingAdapter.PART_DockingManager_Loaded` which calls `LoadDockState` of Syncfusion's `DockingManager`. It takes ~5 000 ms to process and layout the dock. So far I couldn't find an easy fix, while still rendering the dock when no project is open.

FileIcon template takes ~1 000ms to load SVG file. SharpVector dependency allows asynchronous loading but I couldn't make it works smoothly for now. Moreover this loading time, is not THE major bottle neck...

CLI is updated with a middleware to synchronously load hashes and CRUIDs.

Added comments with format `// Improve FCP (~[time] [unit])` to be explicit about these changes and how they are important regarding UX.